work-tracker-servlet/src/main/java/com/deere/isg/worktracker/servlet/LoggerFilter.java
/**
 * Copyright 2018 Deere & Company
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */


package com.deere.isg.worktracker.servlet;

import javax.servlet.FilterChain;
import javax.servlet.ServletException;
import javax.servlet.ServletRequest;
import javax.servlet.ServletResponse;
import javax.servlet.http.HttpServletRequest;
import java.io.IOException;

/**
 * <b>Note:</b> This class is intended to be used only for non-Spring application
 * to log the start of the request.
 */
public class LoggerFilter extends BaseFilter {
    private WorkLogger logger = WorkLogger.getLogger();

    @Override
    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
        if (getOutstanding() != null) {
            logger.logStart((HttpServletRequest) request, getOutstanding().current().orElse(null));
        }
        chain.doFilter(request, response);
    }

    void setLogger(WorkLogger logger) {
        this.logger = logger;
    }
}
